feat: Coupon Entity 구현
- hashtag, description 추가
