+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/CouponType.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.domain;
+package com.woowacourse.kkogkkog.coupon.domain;
 
 import lombok.Getter;
