+++ backend/src/main/java/com/woowacourse/kkogkkog/coupon/domain/repository/CouponRepository.java
@@ -1,0 +1,8 @@
+package com.woowacourse.kkogkkog.coupon.domain.repository;
+
+import com.woowacourse.kkogkkog.coupon.domain.Coupon;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
+
+}
