--- backend/src/main/java/com/woowacourse/kkogkkog/domain/repository/CouponRepository.java
@@ -1,15 +1,0 @@
-package com.woowacourse.kkogkkog.domain.repository;
-
-import com.woowacourse.kkogkkog.domain.Coupon;
-import com.woowacourse.kkogkkog.domain.Member;
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CouponRepository extends JpaRepository<Coupon, Long> {
-
-    List<Coupon> findAllBySender(Member sender);
-
-    List<Coupon> findAllByReceiver(Member receiver);
-}
